Group network-related functionality
