+++ mellow/nn.py
@@ -4,6 +4,9 @@ import jax.numpy as np
 import jax.ops as jo
 
 
+# ------------- network constructors -------------
+
+
 def depth(inp, out, params):
     """Deduces depth of hidden layer."""
     dis = 1 + 4 * inp * (inp - 1) + 4 * out * (out - 1) + 8 * len(params)
@@ -32,3 +35,31 @@ def nd_vect(shape):
     v = np.zeros(int(Σ), dtype=float)
 
     return jo.index_update(v, 0, 1.0)
+
+
+# ------------- activation functions -------------
+
+
+def elu(z):
+    """Computes exponential linear unit."""
+    return z if z > 0 else np.exp(z) - 1.0
+
+
+def relu(z):
+    """Computes rectified linear unit."""
+    return z if z > 0 else 0.0
+
+
+def leaky_relu(z, α=0.2):
+    """Computes leaky rectified linear unit."""
+    return z if z > 0 else α * z
+
+
+def sigmoid(z):
+    """Computes logistic sigmoid."""
+    return 1.0 / (1.0 + np.exp(-z)) if z > 0 else np.exp(z) / (1.0 + np.exp(z))
+
+
+def tanh(z):
+    """Computes hyperbolic tangent."""
+    return np.tanh(z)
